backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
chromadb==0.5.5
sentence-transformers==3.0.1
PyMuPDF==1.24.2
pytesseract==0.3.10
Pillow==10.3.0
python-multipart==0.0.9
numpy==1.26.4
pydantic==2.9.2
sqlalchemy==2.0.35
typing-extensions==4.12.2
